2025.11.24 17:50

+++ source/luametatex/source/luarest/lmtdecodelib.c
@@ -643,7 +643,7 @@ static int pnglib_reduce(lua_State *L)
         int nx = lmt_tointeger(L, 2);
         int ny = lmt_tointeger(L, 3);
         int nz = lmt_tointeger(L, 4);
-        if (nx * ny * nz == size) {
+        if (nx * ny * nz == (int) size) {
             unsigned char *result;
             int dx = nx * nz;
             int mx = dx / 2;
